Misc fixes in dash_client
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@4159 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ include/gpac/internal/mpd.h
@@ -93,7 +93,7 @@ typedef struct
 
 #define GF_MPD_SEGMENT_BASE	\
 	u32 timescale;	\
-	u32 presentation_time_offset;	\
+	u64 presentation_time_offset;	\
 	u32 index_range;	\
 	Bool index_range_exact;	\
 	GF_MPD_URL *initialization_segment;	\
